feat: transition webhook-schema to a sink agnostic design
How webhook-schema was designed and implemented (along with its naming conventions) around the idea of only handling webhooks. This has become a tripping hazard as webhook-schema, ancla and caduceus are transitioning to a sink agnostic architecture.

A lot of code that should be housed in webhook-schema was scattered across argus, ancla, caduceus. This commit addresses this as well.